+++ lib/ui/page/home/page/my_profile/widget/login.dart
@@ -123,7 +123,7 @@ class _UserLoginFieldState extends State<UserLoginField> {
             ),
       label: 'label_login'.l10n,
       hint: widget.login == null ? 'label_login_hint'.l10n : widget.login!.val,
-      canBeEmpty: false,
+      clearable: false,
       subtitle: RichText(
         text: TextSpan(
           children: [